[audio_core] RingBuffer::Allocate returns 2 endpoints
When creating a RingBuffer with Allocate, we return both the readable
and the writable stream interface. This is to enable software to both
read and write from the ring buffer.

When creating a hardware-backed ring-buffer with |Create| we only need
the single endpoint since hardware will be either producing or consuming
audio on the other end.

Test: fx run-test audio_core_unittests
Change-Id: Ie851bc752c75aa82c0bbae5bf81ebc37e3996b8f