[Test] Make mock.sh record arguments correctly
Today, mock.sh stores the command line with which it was invoked and
populates BT_MOCK_ARGS with this information. However, BT_MOCK_ARGS is
always a two-item array with $0 (name of program as invoked) and a
single space-separated string containing all remaining arguments. This
leads to a number of issues:

* Leading and trailing spaces in arguments are lost.
* It's impossible to distinguish between an internal space in an
argument and two space-separated arguments.
* Runs of whitespace within an argument are collapsed into a single
space.
* An unpaired double quotation mark will causes the resulting state
script to fail with a syntax error.

This change repairs this issue by using the printf Bash builtin to
escape arguments being written to the state script. This results in
correctly escape parameters being written to the state script in array
syntax. This allows tests using mock.sh to accurately and reliably
inspect arguments positionally.

Bug: 44766
Test: fx self-test --all
Change-Id: Ia1b6d210739f4151e9b2df9cda1feef411c74cab