[scenic][flatland] Flatland::UnlinkFromParent() impl.
Unlinking a Flatland instance from its parent returns the GraphLinkToken used
to initialize the original link, allowing the process to either link a new
Flatland instance or re-link the old Flatland instance after changes are made.
This functionality will be useful in the future for changing the display a
particular Flatland instance is attached to.

Fixed: 44086

TEST=fx shell runtests -t flatland_unittests

Change-Id: I3003b5683b25cae9e6cc47a5db736d78c73d0b45