[zxtest] Add a way to ignore assertions.
DisableAsserts can be used to ignore asserts from any part of the the current
test. This can be used to execute a function which may assert, in a context
where that assertion is not really unexpected because what is being tested
happens after that helepr function is called.

This is somewhat common for storage integration tests that inject failures
at the bottom of the storage stack, but where an immediate failure is not
important as the test may be verifying that the recovery code works as expected.

Bug: 34116
Change-Id: I1071c7e476ad9497f5060b19892d1aaf0b801373