[docs] Correct instructions for setting clang_prefix
Two issues with the current instructions:

1. The earlier instructions use ${CLANG_DIR} as the destination prefix
where clang is installed, but clang_prefix is actually the bin path
used by the build system. So clang_prefix should be set to
${CLANG_DIR}/bin, not ${CLANG_DIR} alone.

2. The --args argument is a GN command, so the string quoting needs to
be shell-quoted so that it's not consumed by the shell interpreter.

Also, specify to use --no-goma, since a custom clang toolchain is
incompatible with using goma builds.

Change-Id: I1f237dfebad5ec790cd71387dbeddfb7ca2b00da