[system-ota-tests] Add total, paving, and OTA timeouts
This adds total, paving, and OTA timeouts to the OTA test, which allows
it to err out if a particular section of the test gets stuck. This is
needed for longevity and cycle testing to make sure a bad OTA doesn't
sit on a test device until the test executor times out.

Test: Verified timeouts work by hand, and also verified test.json gets
correctly generated with the right flags.

Fixed: 44778

Change-Id: If1d32cdf516eae73d5d6d6e21d587c78be0416cf