[escher] Create render passes without a command buffer.
Reland Step 4 in a series of CLs to build Vulkan pipelines at startup,
rather than lazily as needed.

The main changes are to RenderPassInfo and RenderPassCache.  The
former gains new initialization helpers that don't require images
(images are only needed to obtain a framebuffer while rendering,
but this isn't necessary during warmup).  It also gains additional
validation.

RenderPassCache now features the ability to allow/disallow lazy
creation of render passes.  When disallowed, nullptr will be
returned if no pass is cached.  This behavior can be overridden
with an optional closure.  For example, this allows logging to
Cobalt instead of failing to render a frame.

Testing:
- Modifies FramebufferAllocator to make it more amenable to testing;
  it will now return nullptr instead of DCHECKing if no render-pass
  is available.  New test verifies this, as well as other lazy
  render-pass flows.
- All new RenderPassCache and RenderPassInfo functionality is tested.

Change-Id: I122899b20e23f54e1001520f088023a64a970167