[trace] Replace system-temp with isolated-temp.
The trace_integration_tests utilized system-temp as an IPC mechanism to
pass trace files back to their parent. They also provided a test-temp
mapping that provided a wormhole capability back to the parents /tmp.
This needed to have an extra level of indirection since 3 components are
run, the test, the test app launcher and finally the trace component.
Previously only trace use test-temp. This change maps this all the way
back up to the root so they all write to /test-temp except the actual
test.

Fixes: 25092
Change-Id: I9c01b0ff3fac18a7dc790305da282f8181403458