[build] Set up regime to triage lsan bugs
Firstly, this changes the hack to suppress ASan leak detection
across the board.  Instead of always prepending "detect_leaks=0"
to asan_default_options, it just makes that the default value for
Fuchsia targets when no build argument is given.  The way to
disable the blanket suppression now is:
    asan_default_options = ""
    zircon_extra_args = { asan_default_options = "" }

As leak bugs are found, file a bug #nnnnn for the leaky test/component,
and then add the `suppress-lsan.DO-NOT-USE-THIS` target to `deps` with
a TODO comment referencing that bug.  See the comments above that target,
in //build/config/sanitizers/BUILD.gn for the legacy build and in
//zircon/public/gn/config/instrumentation/BUILD.gn for the Zircon build.

When all those bugs are filed, the `asan_default_options = ""`
default will be reinstated with the individual suppressions keeping
the asan builds' tests all green.  Then each leak bug can be fixed
individually to remove each suppression.  When all are gone, the deps
to support them will be removed too.

Bug: 45047
Change-Id: I4610ca23ebd686d40838061c981abcaccdaabd83