[appmgr] Remove the "system-temp" feature.
This changeset removes the "system-temp" feature which is a big security
win removing global mutable /tmp from Fuchsia. Components can still
utilize "isolated-temp" which will appear as /tmp in their namespace and
will be mapped at one of the following locations in fx shell:

1. /tmp/r/sys/package_name - For components in the system realm.
2. /tmp/r/sys/r/test_env_<random_numbers>/package - For test components.

This change also removes the documentation and tests that are no longer
required. It also invalidates cmx in the schema itself which may cause
build breaking in the petals. If this does break a petal the correct fix
should be to delete "system-temp" from that package and replace it with
"isolated-temp".

Fixes: 25081
Change-Id: I78910553384e6926a84ff015ccdc3237c1e7d386