[netstack] modify dhcp test delays
A recent refactoring changed the delay mechanism here to be synchronous,
causing the delays and timeouts used (over at a ratio of 2:1) to permit
the client to make one optimistic read of the server's reply before
checking that its context was expired; move delay inducement to the
server to ensure the client has time to notice that its context is
expired.

Reduce the overall duration of the test, since this was necessary to
reproduce the flakiness in a reasonable amount of time.

Test:
  fx run-test netstack_dhcp_gotests -- \
  --test.count 10000 \
  --test.failfast \
  --test.run TestDelayRetransmission

Fixes: 45022
Change-Id: If0d0209af41e85127c8748393cbdfd1ce8fbeb16