[llcpp] New table builder that takes tracking_ptr
Some notes:
- The Builder now stores tracking_ptr<Frame> rather than Frame
- To replace the current use cases, we can store the Builder next
to a Frame and have the Builder point to the Frame. However, if
the parent object is moved or copied the Frame will move to a new
location and the Builder's pointer will point to the old location.
To avoid this, the Frame has changed to be non-movable and
non-copyable.
- Several existing use cases depend on moving the existing builder
without significant rework (or switching to allocators when they
are ready to avoid the issue entirely)
- As a result, to make it more practical to deal with the unowned
(no-allocator) case, as second UnownedBuilder was defined which
looks and acts like the original Builder and is movable. This
does not inherit from Builder for two reasons:
  - This would involve complicated logic to keep the Builder
pointer up to date when moved
  - All of the Builder set_* methods return a Builder&& which
 means UnownedBuilder u = UnownedBuilder().set_x(&x).build()
 would be a compiler error.

Test: table_test.cc and GIDL (using Builder not UnownedBuilder)
Test: fx test fidl_llcpp_types_test
Test: fx test fidl_llcpp_conformance_test
Test: fx test fidl_llcpp_builder_test

Change-Id: Ia0d5d582f1f8d595395cf4837685bc1afc2bafd4