[shell] create protocol
This is the bare minimum to be able to test something.

Nothing really useful is implemented yet.

The goal is to be able to establish a connection with an interpreter.

Each new connection creates a new instance within the server. That means
that it's not possible yet to reconnect to an existing instance.

Change-Id: I1a6170b70685a57dec07ea23c53899dfed6eb396