[audio_core] Update MixStage packet frame calculations
When we request subsequent buffers, use a starting frame that aligns
with the end of the previous buffer instead of recalculating the
starting frame value.

This fixes an infinite loop when MixStage is sampling from RingBuffers
and the RingBuffer returns a buffer that is smaller that the mixers
positive filter width. This issue didn't manifest with packet sources
since those streams will naturally seek to the next packet when the
previous is released and the following packet is requested with a
subsequent call to LockBuffer. Since the ring buffer Stream
implementation does not have this implicit seeking behavior, we teach
the MixStage loop to seek itself by explicitly requesting the next frame
from the Stream.

We also take into account the filter width when requesting frames from
the source stream. This fixes a similar issue where we need to keep
requerying the source stream for additional bytes.

Test: fx run-test audio_core_unittests
Change-Id: Ibb0f8bb6a817c6d7697c3dab7200d5de5c80e1df