[stream] Add placeholder for stream in fuchsia.io.FileObject
The fuchsia.io.FileObject struct has only one use, as a possible variant
of the NodeInfo union. While that struct is 4 bytes, it is used in an
aligned context which will add 4 bytes of zeros at the end. We can
therefore safely grow the struct to add a `handle<stream>?` without
breaking the ABI as long as servers do not send any handles in this
field.  Once all the clients have updated to understand this field,
servers can start populating it with an actual handle.

Change-Id: I133a066a7e7a37ea97fe4fc037bb6f271dd4385e