[FDF] When suspending,do not wait for completion of remove-task
At the moment, although we know the device is in the porcess of removal,
we add a dependency on the device's remove task and wait for the device's removal
task to complete. This is unnecessary as the device is going to be
removed anyway and there is no point sending a suspend to a removed
device. This change does not wait on device to be removed and completes
the suspend right away. Also, we will not be adding a suspend task for
child, if we know its being removed.

Bug:45034

Test:fx run-test driver_manager_tests

Change-Id: Ibbf55c8c4cc500e9da677d1ad54e4dd9476449c7