[assertions] Use a debug exception for assertions.
Previously fxl::BreakDebugger used abort() which doesn't currently
provide good backtraces and doesn't allow good contiunuation in a
debugger.

Replace it with a __builtin_debugtrap() which provides a stack and the
stack is more useful because there's no libc code in it. This also
allows stepping past the exception in the debugger.

Chrome uses __builtin_trap(). According to its comments this is because
Crashpad doesn't report debug exceptions. However, on Fuchsia I verified
that crashpad generates dumps for these exceptions.

Bug: 45306
Change-Id: I87b1cce6501a26be1a8e4f4a40b4f355e7ce1fdc