[system-ota-tests] Skip system-update-checker to help land Omaha
Omaha at the moment doesn't have the ability to customize which server
we are talking to, which means once we land it, it would break the OTA
test's ability to talk to a custom repository. For the short term, this
patch has the OTA Test use the N -> N' workflow for N-1 -> N and
N' -> N, where we just tell the system updater to update with an
explicit update package URL.

We should undo this once we've figured out a strategy to customize the
Omaha server for testing purposes (see http://fxb/45156 for that work).

Fixed: 45063
Change-Id: Iaa178c7a6c5e61ca0a8a7828614165010262c420