[exception_broker] Early close of the exception handle.
When an exception crashes, the Exception Broker will try to file a
crash report, for which it generates a minidump image.

In the current code, the exception handle will be laying around until
a connection was made to the crash reporter and the call was made. There
is no need to keep the exception handle around for so long, as we only
need it for the minidump generation. After that, we can close the handle
which will release the exception and kill the process.

Change-Id: Icb916b77b52a1f28166fff65a0de130f19034f63