[debugger] Read/Write watchpoint example
This extends HW Breakpointer to have a test where a R/W watchpoint is
hit repeatedly while a controller is removing it, single stepping and
installing it again. This is to constantly exercise a flow that is
commong when using zxdb.

To run:
fx build-push debug_agent_tests && fx shell -- /pkgfs/packages/debug_agent_tests/0/bin/hw_breakpointer aligned_watchpoints

Change-Id: I04f6e2fc96e76526088064fe5d4dff99a7308ba4