[netstack] rewrite ethernet tx path
Move tx buffer reclamation to the Attach goroutine, replacing a
zx_fifo_read with a condition variable in the WritePacket path as the
means of synchronization, netting one fewer syscall made under lock.
Avoid alloc/free cycles to the arena for tx buffers which can be reused
quickly.

Express Write*Packet* in terms of a vectorized write, permitting better
batching. At the time of writing, upstream doesn't make use of this
unless CapabilitySoftwareGSO is set.

Change-Id: I0c9bf2c2b931c580bb378b5e39d786212f30a766