[fidl][llcpp] Fix BindingRef to only hold weak_ptr.
This fixes a bug where BindingRef holds a shared_ptr to the actual
binding object. This means that a BindingRef can prevent the binding
from being destroyed and the OnUnboundFn from being executed.

Replaced some tests with more effective variants ensuring that binding
destruction is not blocked by a live BindingRef.

Bug: None
Test: runtests -t fidl-async-test-test -r 1000
Change-Id: I0a3d031ca6bd441138c7aec023020e7ec5e2a8a3