[minfs] Add power-failure integration tests.
These tests simulate a failure on the storage stack, like a power
loss or unexpected device removal, and verify that the filesystem is
able to recover from data loss without internal corruption.

Although the test can run in a bot, the main point of the test is to
be user-controllable so that it can perform an exhaustive run of every
test case. The defaut parameters (used for bots) only go through a few
runs to verify that the basic test works.

Bug: 34116

Change-Id: I20931d641d0b3f37386d8c37fcc1f8e974917daa