[wlan][sme] Return latest RSSI in status()
ifconfig calls status() to show the current SSID, BSSID and RSSI. RSSI
is always retrieve from the intial config which is a static value
derived from join scan. So it is not very useful.

Now we can use last_rssi that is updated regularly to give semi realtime
RSSI values.

Fixed: 45286
Test: Included. `fx run-test wlan-sme-tests`
Change-Id: I7fea6cd9b03f02cb4616a54555cf2424b3f89d3a