[soundplayer] replace handle<channel> with fuchsia.io.File
In fuchsia.media.sounds.Player/AddSoundFromFile, the second parameter
represents a file from which a sound is to be read. The type of this
parameter was handle<channel>. As of this CL, the type of that
parameter is now fuchsia.io.File. These are identical, ABI-wise, but
the new definition is more specific.

In addition, AddSoundFromFile now returns the duration of the sound,
in case the client wants to know that.

TEST: added integration test that calls AddSoundFromFile
Change-Id: Ic2ffd80f58a619cdf9335d04898113e9dbbbc0b4