[GN SDK] Add script for publishing packages
This script adds a Fuchsia package to the package server that fserve
runs. Once fpublish has been called, the existing fserve instance or any
future one will include this package and have it available to serve to
an appropriate Fuchsia device.

This script and its test are meant to be run from within a project that
is managed by the GN build system and uses the GN SDK. Use generate.py
to create a new project, then run bin/fpublish.sh or
bin/fpublish-test.sh from that project.

Change-Id: I5b69d264ba972fbe5db148c0902cba18e393c481