[escher] Allow Granite Compute Shaders.
Update the Granite command buffer to be able to dispatch
compute shaders. This requires modifying the command buffer
as well as the pipeline state to be able to manage and cache
compute pipelines and layouts and flush necessary compute
data at the right time.

This CL also modifies the PoseBufferLatching shader and the
other ComputeShader classes to make use of this new workflow.

This is part 1 of several CLs to completely remove the ComputeShader
class and its dependencies from Escher.

Testing: Test to make sure that existing compute shader tests continue
to pass now that they're switched over to using the granite command
buffer, which they do.

SCN-619: Now obsolete/closed

Change-Id: I8748045e0bfdaeb2bb271dcad6891ac6199b6395