[components] Make Start an action
Component start is an activity that should only occur once at a time per
component. Accomplish this by adding a Start action. This also avoids an
edge case where the BeforeStartInstance hook can be triggered twice for
the same start action.

In related refactoring, remove the Clonable trait from model and instead
pass Arc<Model> when the model needs to be cloned.

Change-Id: Iff3b9c985cd7b9bbe145fb87ff315e362e55b14d