[audio][mixer] Apply deadline profiles to mixer threads.
Apply deadline profiles to mixer threads, and adjust how the high/low water
marks are calculated for mixing based on the numbers used in the deadline
profile.  Also add an indication in the AudioOutput::Process trace event of how
close the threads managed to hit their deadline timer.

Tested on Ava using USB audio, and on an Astro using built-in audio.  Verified
that timing targets were being hit using tracing and by monitoring the underflow
reporter in the mixer while under load.

Change-Id: I28da5793f0c3f9409d96125765d416eb0b879686