[build] Add thermd to all board packages
The first of multiple commits in an effort to replace thermd with
PowerManager on Astro. This change adds thermd to the
board_package_labels for every board (should be a no-op because it is
already present in the 'core' product). The next step will be to remove
thermd from the 'core' product.

Change-Id: Ie8b6b52dbcabaab1b3f697b52e92d99406e49985