[kernel][sched] Fix order of generation count update.
Adding the deadline path to QueueThread resulted in the generation
update moving after the run queue insertion operation. This made
it possible for two threads to have the same effective key, when
both have the same start time and stale generation count. The
collision triggers an assert in fbl::WAVLTree.

Fix the order to prevent the possibility of key collisions.

Bug: 45390
Fixed: 45390
Test: System tests + CQ
Change-Id: I343a0ff890737e24086020b5829a5e3692dc8c2b