[media.sessions] Own player control tasks
Prior to this CL, the task to forward control messages to players were
owned by the global executor, requiring cancellation logic to ensure the
clients of the player were disconnected when the player disconnected.

This CL moves ownership of the control task into the player stream
itself. If the player drops, all controlling clients will also drop.

MULTIPLY: `[
  {
    "name": "media_session_tests",
    "total_runs": 100
  }
]`

TEST: fx run-test mediasession_tests
Change-Id: I027b8c2f0be60eb5f34f7fdf694cedc0dd607024