[fidlgen] Extracting HLCPP and libfuzzer backends
Collocating HLCPP templates with the HLCPP codegen, away from the
shared cpp package.

Removing cmdline package: moving flag definition to HLCPP backend and
libfuzzer backend. In a susequent step, we can evolve both backends
indepedently. However, because the HLCPP backend is part of the SDK,
this requires soft migration and migration of all SDK users.

Refactoring LLCPP backend to be closer to what we would want for HLCPP
and libfuzzer. However, refactoring HLCPP and libfuzzer is left for a
susequent step as they both require GN rules rewrite, and will be
easier in isolated CLs.

Test: builds, existing codegen tests
Change-Id: Id5b17ba9dd34c81b7698e53d356e5d0a1ada5112