[components] Make `mark deleting` an action
Currently, a component can be marked deleting in two places: from the
`Realm.DestroyChild` API call, and within the `DeleteChild` action. In
theory, these could run concurrently, causing a duplicate
PreDestroyInstance hook invocation.  Enforce "one-at-a-time" semantics
by moving this code to an action.

Also, make some changes to the `Realm::stop_instance` path so that the
`RealmState` lock is not held the entire time.

Change-Id: I3107b831367c37426a888f2fa7b7dae6f4b68322