[rust][inspect] Disallow duplicate name in test macro
The `assert_inspect_tree` macro only properly matches against a tree
where each property and child has unique name. However, an Inspect
tree may be logged that contains multiple properties or children with
the same name.

As the behavior for reading property or child with non-unique name is
not well defined, this commit updates the test macro so that it throws
an error when it detects such case. The main intention is to provide
warning for users who accidentally log the same name multiple times.

Rustdoc for the macro is also updated to mention that matching duplicate
name now throws an error.

Bug: 42956
Test: New tests added - `fx run-test fuchsia_inspect_tests`

Change-Id: I7ce4e6a9eb2026f8e76ae7fb9436c715aba9a663