[audio_core] allow 3-channel
This CL enables the audio mixer to channel-map between sources and/or
destinations that are 3-channel. This unblocks the usage of 3-channel
input or output devices, and it enables the system to playback or
capture 3-channel streams on 1-, 2- or 3-channel devices. A format
range has been added to virtual_audio_util so that 3-channel virtual
audio devices can be easily created.

This CL also adds unittests for the template-specialized mixer_util
classes that perform the needed format-convert, scale, channel-map and
MAC primitives in the audio processing chain (the template-specialized
SampleNormalizer, SampleScaler, SrcReader and DestMixer classes,
respectively). This CL also adds the first explicit unittests for the
point and linear samplers (these are tested extensively by other
tests, but this scenario was not previously targeted this directly).

Test: adds 22 unit tests to cover the changes

Change-Id: I127644666128537b2747f4dc5db65b5f91f92030