[gn sdk]Adding test_targets.gni generation
This generates a .gni file that contains a template
which defines all the test targets in the GN SDK.
This template is intended for use with build tests
to check the validity of the build rules generated.

The contents of the test_targets.gni was tested by
copying to the samples project and creating an instance of the
template.

Future tests can rely on the test_project which will be part of
CQ.

Change-Id: Icf036156a0c89a2265e06656acca04a48ad15cf8