[netstack] move tcpip/stack.Stack out of the Netstack mutex
Before this change, we maintained our own map of interface id ->
interface state, but now stack.Stack can maintain a reference to
interface state via the Context NICOption (specified when a NIC is
created - cf. stack.CreateNICWithOptions).

All the methods on stack.Stack have internal synchronization, and the
route table and the forwarding table within the stack didn't actually
need to be synchronized for correct operation either - if a client of
netstack modified the routing table, there were no guarantees about when
that modification would affect the data plane (which is not synchronized
via the lock on `type Netstack`) before this change.

other changes:
- use NICInfo.Name where possible
- getNetInterfaces2 now returns all non-address data about a network
interface even if GetMainNICAddress errors
- SetInterfaceStatus now logs errors instead of closing its channel to
the client

Fixed: 43402
Change-Id: Ia884efdfe2b5bd5b850837cf3d6ba18d17aae561