[system_updater] skip pre-flight package installed check
pkg-resolver is just going to repeat the check, so we don't need to
check /pkgfs/versions ourselves.

This does make it a tiny bit slower at the moment; at a device on my
desk, the typical slowdown over 20 runs was ~0.6s/OTA (in the worst case of
nearly no packages needing an update). In the future, we'll hopefully be
able to reduce the affect of the additional latency by doing concurrent
package installs.

Bug: 45129
Change-Id: I75b390a3fa62d40311bdab46e787f2adc9a5d7e7