[files_async] Propagate an error that tells what directory is hanging
There have been *multiple* issues recently about flutter out/
subdirectories hanging. Currently we had a timeout in `iquery` on the
`readdir_recursive` call that would print an error to stderr whenever
the call exceeded the deadline. However, the error logs are quite
useless as we cannot tell what directory is causing the the hang.

With this CL we should be able to tell.

TESTED=fx run-test files_async_tests

Fixed: 45423

Change-Id: Ibb75f71eaa5af7b74b2a0f4037909baabb220bc2