[driver_manager] Don't pass redundant fd 0 to fdio_spawn_etc
fdio_spawn_etc will soon start enforcing that the actions passed
do not include multiple actions with the same target fd number.
driver_manager today violates this constraint.

This is a simple fix that just avoids the redundancy next to the
fdio_spawn_etc call.  It would probably be better to clean things
up in the future so that LaunchWithLoader doesn't have to guess
what its callers passed.

Bug: 45459
Change-Id: Ib20f59841b32f301cba41d6a1979c3409134a6b5