[ramdevice_client] move ramdisk rust lib to //src/lib/storage
moving this library to //src/lib/storage more accurately reflects the
desired repository structure. this library is intended for general use,
not just use within the //src/storage subdirectory.

also, we piggyback a conversion to underscores in the library name onto
this change since we are changing dep lines anyway.

Test: fx run-test ramdevice_client_tests

Change-Id: I6b96af030e4c82f63e5e07dbd09b601f8dfed7b7