[debugger] Remove old symbol test code.
The symbol factory test was written before the index existed so the test
had a complex hardcoded way to find the test functions. This removes
that code and uses the index for the lookup.

Updates the TestSymbolModule to actually load a ModuleSymbols if needed.
Use this in some tests to simplify setup.

Change-Id: I3a03d65175cfcdfcf3357c9e87398be95067e07a