[build] flatten garnet/packages/tests
It is likely many of the individual targets were in use by a small number
of users, but were not referenced anywhere in the build graph. Users are
advised to recreate the necessary groups in an appropriate spot local to
their area instead of globally in garnet/packages. This flattening
drastically simplifies accurate dependency search for incoming
dependencies in the rest of the garnet/packages/* targets.

Change-Id: I67506b69ae0a9cc4a651b7911ffa23de352fa6b5