[kernel] Replace qsort with ktl::stable_sort
Bring std::stable_sort into the ktl subset.  Use it to replace the
one use of qsort in the kernel, which has a comment saying it would
be preferable to use a stable sort anyway.  This makes it possible
to remove qsort from the kernel's minimal C library.

The unused bsearch function is also removed.

Change-Id: I5eb745baf0833f47981d17685481dfc6e49014dc