[GN SDK] Script to connect to a device over SSH
This script finds the Fuchsia device connected to the development host
and connects to it over SSH with appropriate configuration options.

This script and its test are meant to be run from within a project that
is managed by the GN build system and uses the GN SDK. Use generate.py
to create a new project, then run bin/fssh.sh or bin/fssh-test.sh from
that project.

Change-Id: I26ca02e01690c8590b3b0f6aa5bbccb0c9b44cef