[earcons] Fix multiple max earcons after boot
After a reboot, on changing the volume to max, the earcons sound would
play once for every changed volume stream instead of just for media.
This changes it to only play when the media stream is played.

Fixed: 45513
Tests: Added a test that the max volume earcons only plays once if
multiple streams are changed.
fx run-test setui_service_tests -- --test

Change-Id: I9ed2fd40be756dde6eda428c52979af6b0229b82