[fidlc] Add tests to verify that binary OR correctly handles different types.
Previously we had a check in Library::ResolveOrOperatorConstant that
would return an error if the left and right operands of the OR were of
different types. However, this failure was actually never triggered
because binary OR operands would either be converted to the same type
if applicable (in Library::ResolveIdentifierConstant or
Library::ResolveLiteralConstant), or compilation would fail on the
failure of that conversion. So, by the time the OrOperatorConstant was
being resolved, its constituent operands were always of the same type.

Fixed: 43742
Test: OrOperatorDifferentTypesTest in fidlc host test
Change-Id: I10b5b53b6d513c81d43ef9419a755d46a2f944de