[mtd][nand-redundant-storage] Support file backed nand redundant storage
Adds support to treat a file as an mtd device for the mtd-rs tool
Files can be read from or written to as if they were an mtd.

Bytes written to/from a file perfectly match what is written to/from
an mtd so it is possible to pass the same binary blob around between
either mtd backed or file backed interfaces.

Bug: fxb/45053
Test: sudo modprobe nandsim id_bytes=0x2c,0xdc,0x90,0xa6,0x54,0x0 badblocks=5
    cd ~/fuchsia/out/default.zircon/host-x64-linux-clang/obj/system/utest/nand-redundant-storage
    sudo ./nand-redundant-storage-test
Change-Id: I98b08f9e1423cfab0cf72ff149943bedf383e499