[zbi tests] Run on hardware
Note: ZBI tests that specify astro or sherlock will run internally.
In the future there will be a risk of roller failure when a ZBI test is
modified in a way that can pass fuchsia.git CQ but will fail on internal
devices. (This is nothing new though.) For now however, there is no such
risk as astro- and sherlock-targeting ZBI tests will be run in FYI.

Bug: 10198
Change-Id: Ic0ec0ab4301b7544029a827b936a4e693af78cc3