[benchmarking][sl4f] Add system_drm_fps metric
Add a system_drm_fps metric, that implements DRM FPS at the system level
rather than application level.  As is the case with DRM FPS, this metric
is primarily implemented to allow for comparison with other platforms
that implement the same metric.

See newly added comments in "drm_fps.dart" for a more detailed
explanation of the metric.

Change-Id: I6a90cb8d4b351baf09e482ab7ed5fac36f4d56c4