[debugger] Add ambiguous inline query options.
Adds a way for a code block to provide its ambiguous inline call
locations. Querying for functions encounters the same ambiguity that the
Stack currently handles about the ambiguous beginning address of inline
functions.

This function adds a query function for the chain of ambiguous
locations. It is not currently used yet so there should be no behavior
change.

Change-Id: I9096c6038b8c8b93049718f36a498d1f092a59e3