[media.sessions] Implement interruption API
This implements the player interruption API.

Players may specify that, when another higher priority audio stream is
played and reduces the ability of the user to hear their playback, they
should be paused until the user is able to hear them again.

We pause players which are currently playing at the time of an
interruption, and play them again at the end of the interruption.

MULTIPLY: `[
  {
    "name": "mediasession_tests",
    "total_runs": 100
  }
]`

TEST: fx run-test mediasession_tests
Change-Id: I82286ce70232186f8f09625964d3bc765902323b