[audio_core] CreateAudioCapturerWithConfiguration
By taking configuration at creation time, we make it easier for users to
create a working audio capturer, eliminating the complexities of calling
the correct initialization functions in the right order to achieve a
valid state for capture.

This also makes transitions of AudioCapturer types and parameters
easier, because they can take place in a table and xunion instead of
by creating new methods.

Bug: 45240
TEST: no behaviour change
Change-Id: Ia193c6803a5ceb985a451f15ff61c5e8a63db315