[sktool] Modify monolithic message mod.
This splits message into separate mods for message and packet, and as a
consequence also command and util since the command enum and pad
function aren't logically a part of either a message or a packet but are
needed by both.

High line count but there are no functional changes here, only 3
additional tests for the pad function since its now in its own module.

Bug: 44794
Change-Id: I89393ac586cd2342665e7920e713a9c849bd0655