[fidlcat] handle semantic per process
Potentially, two processes could have the same handle numerical value.

This patch compute the handle semantic per process to avoid a potential
collision.

Because we are displaying the handle semantic, all the display functions
must have a way to know the process id for which we are displaying.
This is done by adding a syscall decoder argument to display functions
and by adding a syscall decoder to the printer (the printer is instanciated
to print one syscall and a syscall is associated to only one process id).

Change-Id: Ifba910a199b263806f95f18337d057ef9f88b73f