[build] Change x64 board name from "pc" to "x64"
There are currently two different "board names" for x64: "x64" is
user-facing name and "pc" is the value used by the paver, update tool,
etc. To avoid confusion these names should be consistent.

In particular, the value in the build-info package can be passed in via
a GN argument and this may result in inconsistent board names. This can
cause problems for the update tool which checks the board name value
from the build-info package against the value in the update package.

The paver checks board names, so this CL requires that zedboot be
updated. The version number is bumped accordingly.

Bug: 45562

Change-Id: I944d496cf9eb90de4d25859e4e9b412e82877034