[GN SDK] Add script for running a package server
This script starts a package server that can serve Fuchsia packages to a
Fuchsia device. It also connects to a connected Fuchsia device to
register this package server with that device.

This script and its test are meant to be run from within a project that
is managed by the GN build system and uses the GN SDK. Use generate.py
to create a new project, then run bin/fserve.sh or bin/fserve-test.sh
from that project.

Change-Id: If70343f7cc1ce6fd1256057d871f1868f1877005