[flake] Disable fake_clock_gotests
Fake clock gotests are flaking periodically. This has to do with calls
to `zx_futex_wait` with a finite timeout, the cause of which is still
under investigation. Disabling the tests for now to de-flake.

Bug: 45644
Testing: The disables tests and fixes test flakes. These tests are for
example only anyway, the functionality of fake_clock is tested
elsewhere.

Change-Id: I60d2c302c556da714a9d2cc9296621aaa9e0d089