[components] Remove ambient ELF capability.
This CL removes the ability for components to use the ELF runner by
default. Instead:

  - The ELF runner is offered to the root component under the
    name "elf";

  - Each component is responsible for routing this runner to its
    children if they wish for their children to use the ELF runner;

  - Each component must explicitly state they want to use the ELF
    runner --- it is no longer implicit.

This CL does the minimal change to disable the ambient ELF runner.
More cleanup is required, but to avoid new code being written that
assumes the presence of the ambient ELF runner capability, this
is being checked in now.

Bug: 4761
Change-Id: I6afdfc36c372980c61c55f23f90baeaf080e5e91