[gn sdk] process sysroot atom
Handling for the sysroot sdk atom. This atom contains dist_libs that
are necessary for build configuration so that SDK consumers can target
fuchsia.

Since the dist_libs are the same for each architecture and shouldn't
change names, we reference them directly in build/config/BUILD.gn, so
this handler only copies files and doesn't generate anything.

Change-Id: I5d8fd8038b9af33076e4cc6b0620badd9004d8c5