[fidlcat] add basic semantic for created handles
When a handle is created, we can infer its type. We also add a fd value
to the handle description. This fd value is unique for a given type for
all processes fidlcat is monitoring.

This way the fidlcat output is easier to read.

For example we now have:
zx_handle_close(handle:handle: 5cedf5b7(port:0))

We now know that we are closing a port. The number after "port" is the
fd value and is easier to read than 5cedf5b7.

Change-Id: I54f73e7f88f2912f3d77f8b5034cdd1fa651ff26