[brcmfmac] Enable simulataneous ap & client IFs.
Bug: 29666

Enable simultaneous operation of Client and SoftAP in
the Broadcom driver. As of now, there is only a single
mac address available in Astro and Sherlock platforms.
For this reason, the second interface created (for SoftAP
specifically as the first one is created as a Client at
init time) uses a slightly modified version of the original
mac address.  Here is how the second address is derived
from the original:
- Set bit 0 of byte 0 to 0 (unicast)
- Set bit 1 of byte 0 to 1 (locally administered)
- Increment byte 5 (so it is different from the original)

Broadcom firmware iovar mpc is not recommended in SoftAP mode.
Added code to not set it when SoftAP is enabled.

Test: Manual testing of creating the SoftAP and associating
to it while the client interface associates to an external AP.
Data traffic on both the interfaces has also been verified.
Added a new SIM test to create, start, stop and delete a
SoftAP interface.
Added a SIM tests to associate and disassociate with/from the
SoftAP when a single interface is enabled as well as when
both are enabled.

Change-Id: I1945cbfc511bb99056f8323e4f06880b927320ac