[build] do not add prefixs to prebuilt packages
The prefix as implemented in these rules only applies to the GN side
logic, resulting in update packages and package lists that contain
invalid package names. In order to correctly rename a package, one would
need to create a new package instead. Thoe shortest resolution for now is
to remove the prefix logic.

Change-Id: I4d8dcf4ebc357ebcdd51021b30f6fb59fa64d329