[build] Remove thermd from core.gni
The second in a series of commits to replace thermd with PowerManager on
Astro. This change removes thermd from the 'core' product. This change
should be a no-op because thermd has recently been added to each board
configuration. The next step will be to replace thermd with PowerManager
in the Astro board config.

Change-Id: Ia2d3a489c921b163c260de1db0bfe98c12364d9f