[debugger] Add mockable llvm files.
Originally ModuleSymbolsImpl was a wrapper around the llvm libraries to
query names of things. Over time this interface has gotten fatter and
has more complex logic. Because the code in ModuleSymbolsImpl relies
directly on LLVM, any tests written for it must use real symbol files
loaded by LLVM. This limits the types of tests we can make, especially
for obscure corner cases.

This patch extracts the LLVM code to a new interface. In a subsequent
pass, a mock for this interface will be provided to enable testing more
code.

The abstraction over LLVM is currently very imperfect. As we need to
mock more functionality, we will need to move more code to use our
abstractions instead of reaching into LLVM directly.

Change-Id: If47164b46a035afd32c927077f32f63ee2948fde