[gn-sdk]Add sample project build test
This commit adds a test that takes a generated test project and
attempts to generate and build the Ninja files associated with the
test project. This test project depends on the GN SDK and is intended
to check that a project can be built with the GN SDK. This test will
fail if required arguments (e.g. path to the test project) are
invalid or if the generation or build commands return a non-zero
exit status.

This commit intends to ensure that any changes to Fuchsia's
source code will not break components built with the GN SDK (who's
source code lives outside the Fuchsia source tree).

Change-Id: I968060deb7ee76ed09d5b1d6a88d049a9858bd2e